+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added inline label example for `terra-input-field`, `terra-form-native-select-field`, `terra-form-combobox-field` and `terra-form-single-select-field`.
+  * Added a new example for `terra-list` with interactive elements as a list items.
   * Added new test example for `terra-form-search-select`
   * Added instruction note for invalid example in `terra-search-select` and `terra-combo-box`.
   * Added accessibility note for selectable button in `terra-button`